Factor elliptic solver stack out of XCTS metavars
